Add helper service to make the NTS certificate and key accessible
Chrony reads the certificate and key after dropping privileges, i.e. not
as root. Also, it ignores the supplementary groups of the user it drops
to. This is problematic it the certificate and key are not provided
exclusively for Chrony. The chronyd-cert service helps by providing an
exclusive copy for Chrony.